addressed some feedback, but not yet the iterative invention part
